address PR and move away from own getDamageBonus and use core in all hunter cases
