+++ src/Parser/Hunter/BeastMastery/Modules/Items/Tier19_2p.js
@@ -6,7 +6,7 @@ import PETS from 'common/PETS';
 import SpellIcon from 'common/SpellIcon';
 import SpellLink from 'common/SpellLink';
 import Analyzer from 'Parser/Core/Analyzer';
-import getDamageBonus from 'Parser/Hunter/Shared/Modules/getDamageBonus';
+import getDamageBonus from 'Parser/Core/calculateEffectiveDamage';
 import CorePets from 'Parser/Core/Modules/Pets';
 import ItemDamageDone from 'Interface/Others/ItemDamageDone';
