+++ src/Parser/Hunter/Marksmanship/Modules/Spells/LoneWolf.js
@@ -4,7 +4,7 @@ import Analyzer from 'Parser/Core/Analyzer';
 
 import SPELLS from 'common/SPELLS/index';
 import SpellLink from "common/SpellLink";
-import getDamageBonus from 'Parser/Hunter/Shared/Modules/getDamageBonus';
+import getDamageBonus from 'Parser/Core/calculateEffectiveDamage';
 import ItemDamageDone from 'Interface/Others/ItemDamageDone';
 
 const RAMP_INTERVAL = 6000;
