+++ src/Parser/Hunter/Marksmanship/Modules/Spells/Trueshot.js
@@ -161,7 +161,7 @@ class Trueshot extends Analyzer {
   }
   suggestions(when) {
     when(this.aimedShotThreshold).addSuggestion((suggest, actual, recommended) => {
-      return suggest(<React.Fragment>You only cast {actual} <SpellLink id={SPELLS.AIMED_SHOT.id} />s inside your average <SpellLink id={SPELLS.TRUESHOT.id} /> window. This is your only DPS cooldown, and it's important to maximize it to it's fullest potential by getting as many Aimed Shot squeezed in as possible.</React.Fragment>)
+      return suggest(<React.Fragment>You only cast {actual.toFixed(2)} <SpellLink id={SPELLS.AIMED_SHOT.id} />s inside your average <SpellLink id={SPELLS.TRUESHOT.id} /> window. This is your only DPS cooldown, and it's important to maximize it to it's fullest potential by getting as many Aimed Shot squeezed in as possible.</React.Fragment>)
         .icon(SPELLS.TRUESHOT.icon)
         .actual(`Average of ${actual} Aimed Shots per Trueshot.`)
         .recommended(`>${recommended} is recommended`);
@@ -179,7 +179,7 @@ class Trueshot extends Analyzer {
         .recommended(`15 seconds uptime per cast is recommended`);
     });
     when(this.aimedShotRechargeThreshold).addSuggestion((suggest, actual, recommended) => {
-      return suggest(<React.Fragment>:)</React.Fragment>)
+      return suggest(<React.Fragment>You should make sure to cast <SpellLink id={SPELLS.TRUESHOT.id} /> while you have 0 charges of <SpellLink id={SPELLS.AIMED_SHOT.id} />, to get the most out of the free charge given by activating <SpellLink id={SPELLS.TRUESHOT.id} />.</React.Fragment>)
         .icon(SPELLS.TRUESHOT.icon)
         .actual(`You lost out on ${actual} Aimed Shot charges`)
         .recommended(`${recommended} is recommended`);
