+++ src/Parser/Hunter/Shared/Modules/Items/ZevrimsHunger.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import ITEMS from 'common/ITEMS/index';
 import Analyzer from 'Parser/Core/Analyzer';
-import getDamageBonus from 'Parser/Hunter/Shared/Modules/getDamageBonus';
+import getDamageBonus from 'Parser/Core/calculateEffectiveDamage';
 import ItemDamageDone from 'Interface/Others/ItemDamageDone';
 
 /**
