--- src/Parser/Hunter/Shared/Modules/getDamageBonus.js
@@ -1,4 +1,0 @@
-export default function getDamageBonus(event, increase) {
-  const raw = (event.amount || 0) + (event.absorbed || 0);
-  return raw - (raw / (1 + increase));
-}
