+++ src/Parser/Hunter/Survival/Modules/Items/Tier21_2p.js
@@ -7,7 +7,7 @@ import { formatPercentage } from 'common/format';
 import Analyzer from 'Parser/Core/Analyzer';
 import ItemDamageDone from 'Interface/Others/ItemDamageDone';
 import HIT_TYPES from 'Parser/Core/HIT_TYPES';
-import getDamageBonus from 'Parser/Hunter/Shared/Modules/getDamageBonus';
+import getDamageBonus from 'Parser/Core/calculateEffectiveDamage';
 
 //It's a 100% increase to critical strike damage, which is already a 100% modifier, making it effectively only hit 50% harder than a crit otherwise would.
 const T21_2P_CRIT_DMG_BONUS = 0.5;
